Don't try to use JIT code if we need dex pc move events
We try and transition to JITed code when we start in interpreter if
don't need to run in interpreter. We check for a few conditions but
missed checking if we need to call dex pc move events for this methods.
This CL extends the condition to check if we need dex pc move events.
In practice, this isn't a problem since we are only checking if the
entrypoint is JITed code but impacts only jit-on-first-use config. It is
cleaner to add this check.

Bug: 315751006
Test: art/test.py
Change-Id: I7734eb581dba39ed57c0068d192d481899caea5e