Reland 'Pass functors as rvalues when possible'
On local compiles I saw that DeleteAllImpureWhich was the third
most time consuming method, using pprofs sorting in bottom-up.
By passing the functor it uses as rvalue, the method speeds up
~10% and it drops from the third most time consuming to the fourth.

The other modified methods in the CL are not showing up in the
pprof profile that I took, but changing them to use rvalues
shouldn't affect them negatively.

Test: Locally compile, take a trace, and observe time spent
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Change-Id: I6c363d5601fd4865f4e7881e64b883bd6bbedb69